feat: add auth middleware to protect the routes

+++ backend/src/controllers/notes.ts
@@ -2,12 +2,17 @@ import { RequestHandler } from 'express';
 import { NoteModel } from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../utils/assertIsDefined';
 
 export const getNotesController: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
     // throw Error('Oops!');
     // throw createHttpError(401);
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({
+      userId: authenticatedUserId,
+    }).exec();
     res.status(200).json(notes);
   } catch (err) {
     next(err);
@@ -17,7 +22,11 @@ export const getNotesController: RequestHandler = async (req, res, next) => {
 export const getNoteByIdController: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
 
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
     }
@@ -26,6 +35,11 @@ export const getNoteByIdController: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, 'Note not found');
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, 'You are not authorized to access this note');
+    }
+
     res.status(200).json(note);
   } catch (err) {
     next(err);
@@ -44,12 +58,18 @@ export const createNoteController: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, 'Title is required');
     }
-    const newNote = await NoteModel.create({ title, text });
+    const newNote = await NoteModel.create({
+      title,
+      text,
+      userId: authenticatedUserId,
+    });
 
     res.status(201).json(newNote);
   } catch (err) {
@@ -73,15 +93,26 @@ export const updateNoteController: RequestHandler<
 > = async (req, res, next) => {
   const noteId = req.params.noteId;
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
+    }
+
+    if (!title) {
+      throw createHttpError(400, 'Missing Title fields');
     }
 
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, 'You are not authorized to access this note');
     }
 
     if (title) {
@@ -109,16 +140,26 @@ export const deleteNoteController: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
     }
 
-    const note = await NoteModel.findByIdAndDelete(noteId).exec();
+    const note = await NoteModel.findById(noteId).exec();
+
     if (!note) {
       throw createHttpError(404, 'Note not found');
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, 'You are not authorized to access this note');
+    }
+
+    await NoteModel.findByIdAndDelete(noteId).exec();
 
     res.sendStatus(204);
   } catch (err) {
